requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
click==8.1.7
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
Mako==1.3.5
MarkupSafe==2.1.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
sqlmodel==0.0.22
starlette==0.38.3
typing_extensions==4.12.2
uvicorn==0.30.6
